+++ src/Ritten/Program.cs
@@ -6,6 +6,8 @@ using Ritten.GitHub;
 using Ritten.Workflows.DotNet;
 using Ritten.Workflows.DotNetPackage;
 using Ritten.Workflows.DotNetTool;
+using Wolfe.CommandLine;
+using Wolfe.CommandLine.Completions;
 
 var builder = WorkflowApplication.CreateBuilder();
 
@@ -23,7 +25,9 @@ if (built.IsError)
     return ExitCode.ConfigurationError;
 }
 
-var root = new RootCommand("The Ritten build workflow.");
+var root = new RootCommand("The Ritten build workflow.")
+    .AddCompletions("ritten");
 await root.InstallRitten(built.Value);
+await CompletionAutoInstall.Run("ritten", args);
 
 return await root.Parse(args).InvokeAsync();